src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.1
django-debug-toolbar==3.2.4
django-redis==5.2.0
fire==0.4.0
frozenlist==1.2.0
gql==3.0.0a5
graphql-core==3.1.7
gunicorn==20.1.0
idna==2.10
multidict==5.2.0
mypy-extensions==0.4.3
packageurl-python==0.9.6
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
promise==2.3
psycopg2-binary==2.9.3
pybraries==0.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
regex==2021.11.10
requests==2.27.1
Rx==3.2.0
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.1
typing-extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
websockets==10.1
wrapt==1.13.3
yarl==1.7.2
